flask_demo/requirements/common.txt
alembic==1.13.1
bleach==4.1.0
blinker==1.8.2
click==8.1.7
dominate==2.9.1
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==4.8.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-Moment==1.0.6
Flask-PageDown==0.4.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
html5lib==1.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
Markdown==3.4.1
MarkupSafe==2.1.5
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
python-editor==1.0.4
six==1.16.0
SQLAlchemy==2.0.30
visitor==0.1.3
webencodings==0.5.1
Werkzeug==3.0.3
WTForms==3.1.2
